pyproject.toml
[tool.poetry]
name = "eieio"
version = "0.1.2"
description = "Extensible Imaging Equipment I/O (EIEIO)"
license = "BSD-3-Clause"
authors = [ "Joseph Goldstone <jgoldstone@arri.com>" ]
maintainers = [ "Joseph Goldstone <jgoldstone@arri.com>" ]
# readme = 'README.rst'
keywords = [
    "eieio",
    "data",
    "dataset",
    "datasets",
    "python",
    "spectral-data",
    "spectral-dataset",
    "spectral-datasets"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development"
]

[tool.poetry.dependencies]
python = "^3.7"
imageio = "*"
colour = { path = "/usr/local/repos/git/colour-science/colour", develop = true }
colour-optional-dependencies = { path = "/usr/local/repos/git/colour-science/hacks/colour-optional-dependencies" }
colour-datasets = { path = "/usr/local/repos/git/colour-science/colour-datasets", develop = true }
# grpc = "*"
pyserial = "^3.5"
grpcio = "1.21.1"
grpcio-tools = "1.21.1"
protobuf = "3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = [ "poetry>=0.12" ]
build-backend = "poetry.masonry.api"
